Issue-677: MovieDetailsRepository created
