+++ screen/movie/src/main/java/com/kirchhoff/movies/screen/movie/repository/MovieRepository.kt
@@ -6,11 +6,6 @@ import com.kirchhoff.movies.core.data.UIMovie
 import com.kirchhoff.movies.core.repository.BaseRepository
 import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
-import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.networkdata.core.NetworkPaginated
-import com.kirchhoff.movies.networkdata.details.movie.NetworkMovieDetails
-import com.kirchhoff.movies.networkdata.details.movie.NetworkTrailersList
-import com.kirchhoff.movies.networkdata.main.NetworkMovie
 import com.kirchhoff.movies.screen.movie.mapper.IMovieDetailsMapper
 import com.kirchhoff.movies.screen.movie.mapper.IMovieListMapper
 import com.kirchhoff.movies.screen.movie.network.MovieService
@@ -19,13 +14,9 @@ import com.kirchhoff.movies.storage.movie.IStorageMovie
 
 internal interface IMovieRepository {
     suspend fun fetchDiscoverList(page: Int): Result<UIPaginated<UIMovie>>
-    suspend fun fetchDetails(id: MovieId): Result<NetworkMovieDetails>
     suspend fun similarMovies(id: MovieId, page: Int): Result<UIPaginated<UIMovie>>
-    suspend fun fetchSimilarMovies(id: MovieId, page: Int): Result<NetworkPaginated<NetworkMovie>>
     suspend fun fetchByCountry(countryId: String, page: Int): Result<UIPaginated<UIMovie>>
     suspend fun fetchByCompany(companyId: String, page: Int): Result<UIPaginated<UIMovie>>
-    suspend fun fetchTrailersList(id: MovieId): Result<NetworkTrailersList>
-    suspend fun fetchMovieCredits(id: MovieId): Result<NetworkEntertainmentCredits>
     suspend fun fetchByGenre(genre: String, page: Int): Result<UIPaginated<UIMovie>>
     suspend fun fetchImages(id: MovieId): Result<List<UIImage>>
 }
@@ -48,19 +39,12 @@ internal class MovieRepository(
         return movieListMapper.createMovieList(result)
     }
 
-    override suspend fun fetchDetails(id: MovieId): Result<NetworkMovieDetails> = apiCall { movieService.fetchDetails(id.value) }
-
     override suspend fun similarMovies(id: MovieId, page: Int): Result<UIPaginated<UIMovie>> =
         movieListMapper.createMovieList(
             apiCall {
                 movieService.fetchSimilarMovies(id.value, page)
             }
         )
-
-    override suspend fun fetchSimilarMovies(id: MovieId, page: Int): Result<NetworkPaginated<NetworkMovie>> =
-        apiCall {
-            movieService.fetchSimilarMovies(id.value, page)
-        }
 
     override suspend fun fetchByCountry(countryId: String, page: Int): Result<UIPaginated<UIMovie>> =
         movieListMapper.createMovieList(
@@ -75,18 +59,6 @@ internal class MovieRepository(
                 movieService.fetchByCompany(companyId, page)
             }
         )
-
-    override suspend fun fetchTrailersList(id: MovieId): Result<NetworkTrailersList> = apiCall { movieService.fetchTrailersList(id.value) }
-
-    override suspend fun fetchMovieCredits(id: MovieId): Result<NetworkEntertainmentCredits> {
-        val result = apiCall { movieService.fetchMovieCredits(id.value) }
-
-        if (result is Result.Success) {
-            movieStorage.updateCredits(id.value, result.data)
-        }
-
-        return result
-    }
 
     override suspend fun fetchByGenre(genre: String, page: Int): Result<UIPaginated<UIMovie>> =
         movieListMapper.createMovieList(
