+++ screen/movie/src/main/java/com/kirchhoff/movies/screen/movie/ui/screen/details/MovieDetailsModule.kt
@@ -1,6 +1,8 @@
 package com.kirchhoff.movies.screen.movie.ui.screen.details
 
 import com.kirchhoff.movies.core.data.UIMovie
+import com.kirchhoff.movies.screen.movie.repository.IMovieDetailsRepository
+import com.kirchhoff.movies.screen.movie.repository.MovieDetailsRepository
 import com.kirchhoff.movies.screen.movie.ui.screen.details.usecase.IMovieDetailsUseCase
 import com.kirchhoff.movies.screen.movie.ui.screen.details.usecase.MovieDetailsUseCase
 import com.kirchhoff.movies.screen.movie.ui.screen.details.viewmodel.MovieDetailsViewModel
@@ -9,9 +11,17 @@ import org.koin.dsl.module
 
 internal val movieDetailsModule = module {
 
+    single<IMovieDetailsRepository> {
+        MovieDetailsRepository(
+            movieService = get(),
+            movieStorage = get()
+        )
+    }
+
     single<IMovieDetailsUseCase> {
         MovieDetailsUseCase(
             movieRepository = get(),
+            movieDetailsRepository = get(),
             movieDetailsMapper = get(),
             movieListMapper = get()
         )
